requirements.txt
Flask==2.3.2
Flask-Dance==5.0.0
python-dotenv==1.0.0
gunicorn==20.1.0
